+++ packages/powersync-sdk-common/src/client/sync/stream/AbstractRemote.ts
@@ -7,6 +7,7 @@ import ndjsonStream from 'can-ndjson-stream';
 import { RSocketConnector } from 'rsocket-core';
 import { WebsocketClientTransport } from 'rsocket-websocket-client';
 import { serialize, deserialize } from 'bson';
+import { AbortOperation } from '../../../utils/AbortOperation';
 
 export type RemoteConnector = {
   fetchCredentials: () => Promise<PowerSyncCredentials | null>;
@@ -226,14 +227,46 @@ export abstract class AbstractRemote {
 
     const request = await this.buildRequest(path);
 
+    /**
+     * This abort controller will abort pending fetch requests.
+     * If the request has resolved, it will be used to close the readable stream.
+     * Which will cancel the network request.
+     *
+     * This nested controller is required since:
+     *  Aborting the active fetch request while it is being consumed seems to throw
+     *  an unhandled exception on the window level.
+     */
+    const controller = new AbortController();
+    let requestResolved = false;
+    abortSignal?.addEventListener('abort', () => {
+      if (!requestResolved) {
+        // Only abort via the abort controller if the request has not resolved yet
+        controller.abort(
+          abortSignal.reason ??
+            new AbortOperation('Cancelling network request before it resolves. Abort signal has been received.')
+        );
+      }
+    });
+
     const res = await fetch(request.url, {
       method: 'POST',
       headers: { ...headers, ...request.headers },
       body: JSON.stringify(data),
-      signal: abortSignal,
+      signal: controller.signal,
       cache: 'no-store',
       ...options.fetchOptions
+    }).catch((ex) => {
+      if (ex.name == 'AbortError') {
+        throw new AbortOperation(`Pending fetch request to ${request.url} has been aborted.`);
+      }
+      throw ex;
     });
+
+    if (!res) {
+      throw new Error('Fetch request was aborted');
+    }
+
+    requestResolved = true;
 
     if (!res.ok || !res.body) {
       const text = await res.text();
@@ -243,8 +276,55 @@ export abstract class AbstractRemote {
       throw error;
     }
 
-    // TODO handle closing errors better like on main
-    const jsonS = ndjsonStream(res.body!);
+    /**
+     * The can-ndjson-stream does not handle aborted streams well.
+     * This will intercept the readable stream and close the stream if
+     * aborted.
+     */
+    const reader = res.body.getReader();
+    // This will close the network request and read stream
+    const closeReader = async () => {
+      try {
+        await reader.cancel();
+      } catch (ex) {
+        // an error will throw if the reader hasn't been used yet
+      }
+      reader.releaseLock();
+    };
+
+    abortSignal?.addEventListener('abort', () => {
+      closeReader();
+    });
+
+    const outputStream = new ReadableStream({
+      start: (controller) => {
+        const processStream = async () => {
+          while (!abortSignal?.aborted) {
+            try {
+              const { done, value } = await reader.read();
+              // When no more data needs to be consumed, close the stream
+              if (done) {
+                break;
+              }
+              // Enqueue the next data chunk into our target stream
+              controller.enqueue(value);
+            } catch (ex) {
+              this.logger.error('Caught exception when reading sync stream', ex);
+              break;
+            }
+          }
+          if (!abortSignal?.aborted) {
+            // Close the downstream readable stream
+            await closeReader();
+          }
+          controller.close();
+        };
+        processStream();
+      }
+    });
+
+    const jsonS = ndjsonStream(new Response(outputStream).body);
+
     const stream = new DataStream({
       logger: this.logger
     });
@@ -253,16 +333,22 @@ export abstract class AbstractRemote {
 
     const l = stream.registerListener({
       lowWater: async () => {
-        const { done, value } = await r.read();
-        // Exit if we're done
-        if (done) {
+        try {
+          const { done, value } = await r.read();
+          // Exit if we're done
+          if (done) {
+            stream.close();
+            l?.();
+            return;
+          }
+          stream.enqueueData(value);
+        } catch (ex) {
           stream.close();
-          l?.();
-          return;
+          throw ex;
         }
-        stream.enqueueData(value);
       },
       closed: () => {
+        closeReader();
         l?.();
       }
     });