+++ packages/powersync-sdk-common/src/client/sync/stream/AbstractStreamingSyncImplementation.ts
@@ -15,6 +15,7 @@ import { BucketChecksum, BucketStorageAdapter, Checkpoint } from '../bucket/Buck
 import { SyncStatus, SyncStatusOptions } from '../../../db/crud/SyncStatus';
 import { SyncDataBucket } from '../bucket/SyncDataBucket';
 import { BaseObserver, BaseListener, Disposable } from '../../../utils/BaseObserver';
+import { AbortOperation } from '../../../utils/AbortOperation';
 
 export enum LockType {
   CRUD = 'crud',
@@ -50,6 +51,14 @@ export interface AbstractStreamingSyncImplementationOptions {
 }
 
 export interface StreamingSyncImplementationListener extends BaseListener {
+  /**
+   * Triggered whenever a status update has been attempted to be made or
+   * refreshed.
+   */
+  statusUpdated?: ((statusUpdate: SyncStatusOptions) => void) | undefined;
+  /**
+   * Triggers whenever the status' members have changed in value
+   */
   statusChanged?: ((status: SyncStatus) => void) | undefined;
 }
 
@@ -90,6 +99,7 @@ export abstract class AbstractStreamingSyncImplementation
   protected options: AbstractStreamingSyncImplementationOptions;
   protected abortController: AbortController | null;
   protected crudUpdateListener?: () => void;
+  protected streamingSyncPromise?: Promise<void>;
 
   syncStatus: SyncStatus;
   triggerCrudUpload: () => void;
@@ -229,16 +239,53 @@ export abstract class AbstractStreamingSyncImplementation
     if (this.abortController) {
       await this.disconnect();
     }
+
     this.abortController = new AbortController();
-    this.streamingSync(this.abortController.signal);
-    return this.waitForStatus({ connected: true });
+    this.streamingSyncPromise = this.streamingSync(this.abortController.signal);
+
+    // Return a promise that resolves when the connection status is updated
+    return new Promise<void>((resolve) => {
+      const l = this.registerListener({
+        statusUpdated: (update) => {
+          // This is triggered as soon as a connection is read from
+          if (typeof update.connected == 'undefined') {
+            // only concern with connection updates
+            return;
+          }
+
+          if (update.connected == false) {
+            /**
+             * This function does not reject if initial connect attempt failed
+             */
+            this.logger.warn('Initial connect attempt did not successfully connect to server');
+          }
+
+          resolve();
+          l();
+        }
+      });
+    });
   }
 
   async disconnect(): Promise<void> {
     if (!this.abortController) {
-      throw new Error('Disconnect not possible');
+      return;
     }
-    this.abortController.abort('Disconnected');
+
+    // This might be called multiple times
+    if (!this.abortController.signal.aborted) {
+      this.abortController.abort(new AbortOperation('Disconnect has been requested'));
+    }
+
+    // Await any pending operations before completing the disconnect operation
+    try {
+      await this.streamingSyncPromise;
+    } catch (ex) {
+      // The operation might have failed, all we care about is if it has completed
+      this.logger.warn(ex);
+    }
+    this.streamingSyncPromise = undefined;
+
     this.abortController = null;
     this.updateSyncStatus({ connected: false });
   }
@@ -266,7 +313,11 @@ export abstract class AbstractStreamingSyncImplementation
     let nestedAbortController = new AbortController();
 
     signal.addEventListener('abort', () => {
-      nestedAbortController.abort();
+      /**
+       * A request for disconnect was received upstream. Relay the request
+       * to the nested abort controller.
+       */
+      nestedAbortController.abort(signal?.reason ?? new AbortOperation('Received command to disconnect from upstream'));
       this.crudUpdateListener?.();
       this.crudUpdateListener = undefined;
       this.updateSyncStatus({
@@ -277,6 +328,12 @@ export abstract class AbstractStreamingSyncImplementation
       });
     });
 
+    /**
+     * This loops runs until [retry] is false or the abort signal is set to aborted.
+     * Aborting the nestedAbortController will:
+     *  - Abort any pending fetch requests
+     *  - Close any sync stream ReadableStreams (which will also close any established network requests)
+     */
     while (true) {
       try {
         if (signal?.aborted) {
@@ -284,27 +341,45 @@ export abstract class AbstractStreamingSyncImplementation
         }
         const { retry } = await this.streamingSyncIteration(nestedAbortController.signal);
         if (!retry) {
+          /**
+           * A sync error ocurred that we cannot recover from here.
+           * This loop must terminate.
+           * The nestedAbortController will close any open network requests and streams below.
+           */
           break;
         }
         // Continue immediately
       } catch (ex) {
-        this.logger.error(ex);
+        /**
+         * Either:
+         *  - A network request failed with a failed connection or not OKAY response code.
+         *  - There was a sync processing error.
+         * This loop will retry.
+         * The nested abort controller will cleanup any open network requests and streams.
+         * The WebRemote should only abort pending fetch requests or close active Readable streams.
+         */
+        if (ex instanceof AbortOperation) {
+          this.logger.warn(ex);
+        } else {
+          this.logger.error(ex);
+        }
+        await this.delayRetry();
+      } finally {
+        if (!signal.aborted) {
+          nestedAbortController.abort(new AbortOperation('Closing sync stream network requests before retry.'));
+          nestedAbortController = new AbortController();
+        }
+
         this.updateSyncStatus({
           connected: false
         });
-        // On error, a little before retrying
-        await this.delayRetry();
-      } finally {
-        // Abort any open network requests. Create a new nested controller for retry.
-        nestedAbortController.abort();
-        nestedAbortController = new AbortController();
+
+        // On error, wait a little before retrying
       }
     }
 
     // Mark as disconnected if here
-    if (this.abortController) {
-      await this.disconnect();
-    }
+    this.updateSyncStatus({ connected: false });
   }
 
   protected async streamingSyncIteration(signal: AbortSignal, progress?: () => void): Promise<{ retry?: boolean }> {
@@ -503,8 +578,12 @@ export abstract class AbstractStreamingSyncImplementation
 
     if (!this.syncStatus.isEqual(updatedStatus)) {
       this.syncStatus = updatedStatus;
+      // Only trigger this is there was a change
       this.iterateListeners((cb) => cb.statusChanged?.(updatedStatus));
     }
+
+    // trigger this for all updates
+    this.iterateListeners((cb) => cb.statusUpdated?.(options));
   }
 
   private async delayRetry() {