+++ packages/powersync-sdk-common/src/index.ts
@@ -27,6 +27,7 @@ export * from './db/DBAdapter';
 export * from './db/Column';
 export * from './db/schema/TableV2';
 
+export * from './utils/AbortOperation';
 export * from './utils/BaseObserver';
 export * from './utils/strings';
 export * from './utils/DataStream';