+++ demos/react-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,14 @@
 # react-supabase-todolist
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [6c43ec6]
+- Updated dependencies [8f7caa5]
+  - @journeyapps/powersync-sdk-web@0.4.0
+  - @journeyapps/powersync-react@1.1.1
+
 ## 0.2.2
 
 ### Patch Changes